==> notebooks/02_cleaning1.ipynb <==
import pandas as pd
import numpy as np

# Check if running in Google Colab
if 'google.colab' in str(get_ipython()):
    from google.colab import drive
    drive.mount('/content/drive')

    # Define the base path for your data directory for Colab
    basePath = '/content/drive/MyDrive'

else:
    # Running locally or in a different environment
    basePath = '..'

file_path = 'data/processed/unified_countries.csv'

df = pd.read_csv(f'{basePath}/{file_path}')
df.head(3)

# 3. Remove duplicate columns
df = df.loc[:, ~df.columns.duplicated()]

non_numeric_cols = ['STATION', 'NAME', 'DATE', 'Source_Station_ID', 'Source_Country_Name', 'Source_Station_Name']
numeric_cols = [col for col in df.columns if col not in non_numeric_cols]

df[numeric_cols] = df[numeric_cols].apply(pd.to_numeric, errors='coerce')

# 5. Rearrange just for better look
reordered_cols = (
    ['Source_Country_Name', 'Source_Station_ID', 'Source_Station_Name',
     'STATION', 'NAME', 'DATE', 'LATITUDE', 'LONGITUDE', 'ELEVATION']
    + sorted([col for col in df.columns if col not in [
        'Source_Country_Name', 'Source_Station_ID', 'Source_Station_Name',
        'STATION', 'NAME', 'DATE', 'LATITUDE', 'LONGITUDE', 'ELEVATION']])
)
df = df[reordered_cols]

df.head(3)

# Renaming for clarity first
df = df.rename(columns={
    'Source_Country_Name': 'COUNTRY',
    'NAME': 'STATION_NAME'
})

df = df.drop(columns=['Source_Station_ID', 'Source_Station_Name'])

threshold = 0.8
missing_ratio = df.isnull().mean()
df = df.loc[:, missing_ratio <= threshold]

output_path = 'data/processed/cleaned_data1.csv'

df.to_csv(f'{basePath}/{output_path}', index=False)

print(f"Cleaning complete. Saved as '{output_path}'")
