==> notebooks/05_feature1.ipynb <==
# Check if running in Google Colab
if 'google.colab' in str(get_ipython()):
    from google.colab import drive
    drive.mount('/content/drive')

    # Define the base path for your data directory for Colab
    basePath = '/content/drive/MyDrive'

else:
    # Running locally or in a different environment
    basePath = '..'

import pandas as pd
import numpy as np

df_grouped = pd.read_csv(f"{basePath}/data/processed/grouped_data1.csv")
df_grouped

#Checking if how many countries
df_grouped['COUNTRY'].nunique()

# Get the set of unique years for each country
years_by_country = df_grouped.groupby('COUNTRY')['DATE'].apply(set)

common_years = set.intersection(*years_by_country)

common_years_list = sorted(list(common_years))

print("Years present in all countries:")
print(common_years_list)

if common_years_list:
    oldest_common_year = common_years_list[0]
    print(f"\nThe oldest year present in all countries is: {oldest_common_year}")
else:
    print("\nThere are no years that are present in all countries.")

#Now since that no year that is present in all of the countries...let's see if 90% countries have the common date

total_countries = df_grouped['COUNTRY'].nunique()

country_counts_by_year = df_grouped.groupby('DATE')['COUNTRY'].nunique()

percentage_countries_by_year = (country_counts_by_year / total_countries) * 100

# Define the threshold (90%)
percentage_threshold = 90

# Find the years where the percentage of countries is above the threshold
years_with_sufficient_data = percentage_countries_by_year[percentage_countries_by_year >= percentage_threshold].index.tolist()

print(f"Years present in at least {percentage_threshold}% of countries:")
print(years_with_sufficient_data)

# Filter the grouped DataFrame to keep only the countries that have data in any of the years with sufficient data
if years_with_sufficient_data:
    countries_to_keep = df_grouped[df_grouped['DATE'].isin(years_with_sufficient_data)]['COUNTRY'].unique()
    df_filtered = df_grouped[df_grouped['COUNTRY'].isin(countries_to_keep)].copy()
    print("\nFiltered DataFrame containing countries with data in the identified common years:")
    display(df_filtered)
    print(f"\nNumber of countries in the filtered DataFrame: {df_filtered['COUNTRY'].nunique()}")
else:
    df_filtered = pd.DataFrame() 
    print("\nNo years found that are present in at least 90% of countries. The filtered DataFrame is empty.")
